requirements.txt
openai==1.11.0
langchain==0.1.5
langchain-community==0.0.17
langchain-openai==0.0.5
streamlit==1.31.0
streamlit-authenticator==0.3.1
streamlit-aggrid==0.3.4.post3
matplotlib==3.8.2
plotly==5.18.0
scipy==1.11.4
scikit-learn==1.3.2
transformers==4.36.2
python-dotenv==1.0.0
azure-ai-documentintelligence==1.0.0b1
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-ai-contentsafety==1.0.0b1
requests==2.31.0
tiktoken==0.5.2
beautifulsoup4==4.12.2
fake-useragent==1.4.0
chardet==5.2.0
azure-search-documents==11.4.0
opencensus-ext-azure==1.1.13
pandas==2.2.0
python-docx==0.8.11
openpyxl==3.1.2
xlsxwriter==3.1.9